Added subnavbar within Gereate quest based off if the person is famous or not
