+++ website/client/src/App.js
@@ -3,15 +3,19 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 import NavBar from "./comp/NavBar";
 import About from "./comp/about";
-import Prompt from "./comp/prompt";
+import Generatequest from "./comp/Generatequest";
+import FamousPerson from "./comp/Famousperson";
+import HonoredOne from "./comp/Honoredone";
 
 function App() {
     return (
         <Router>
             <NavBar />
             <Routes>
-                <Route path="/prompt" element={<Prompt />} />
+                <Route path="/generate-quest" element={<Generatequest />} />
                 <Route path="/about" element={<About />} />
+                <Route path="/famous-person" element={<FamousPerson />} />
+                <Route path="/honored-one" element={<HonoredOne />} />
             </Routes>
         </Router>
     );
