+++ website/client/src/comp/Famousperson.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import SubNavBar from './SubNavBar';
+
+
+function FamousPerson() {
+    
+    return (
+        <div>
+            <SubNavBar />
+            <h1>Famous Person</h1>
+        </div>
+    );
+}
+
+export default FamousPerson;
