+++ website/client/src/comp/Honoredone.js
@@ -1,0 +1,14 @@
+import React from "react";
+import SubNavBar from "./SubNavBar";
+
+function HonoredOne() {
+
+    return (
+        <div>
+            <SubNavBar />
+            <h1>Honored One</h1>
+        </div>
+    );
+}
+
+export default HonoredOne;
